[MIG] stock_card_report: Migration to 15.0
