+++ stock_card_report/__manifest__.py
@@ -4,7 +4,7 @@
 {
     "name": "Stock Card Report",
     "summary": "Add stock card report on Inventory Reporting.",
-    "version": "14.0.1.0.2",
+    "version": "15.0.1.0.0",
     "category": "Warehouse",
     "website": "https://github.com/OCA/stock-logistics-reporting",
     "author": "Ecosoft, Odoo Community Association (OCA)",
@@ -17,5 +17,11 @@
         "reports/stock_card_report.xml",
         "wizard/stock_card_report_wizard_view.xml",
     ],
+    "assets": {
+        "web.assets_backend": [
+            "stock_card_report/static/src/css/**/*",
+            "stock_card_report/static/src/js/**/*",
+        ]
+    },
     "installable": True,
 }
