+++ stock_card_report/readme/CONTRIBUTORS.rst
@@ -1,2 +1,3 @@
 * Pimolnat Suntian <pimolnats@ecosoft.co.th>
 * Prapassorn Sornkaew <prapassorn.s@prothaitechnology.com>
+* Emad Shaaban <emad.shaaban92@gmail.com>
